Assignment-3 modifications
