+++ assignment-3/src/components/InputFile.js
@@ -1,0 +1,9 @@
+const InputFile = () => {
+    return (
+      <div style = {{ marginTop: "20px" }}>
+        <input type = "file" />
+      </div>
+    );
+}
+
+export default InputFile;
