+++ assignment-3/src/components/InputText.js
@@ -1,0 +1,5 @@
+const InputText = () => {
+    return <input type = "text" placeholder = "Enter Text Here..." />
+}
+
+export default InputText;
