+++ assignment-3/src/components/SelectOptions.js
@@ -1,0 +1,11 @@
+const SelectOptions = () => {
+    return (
+      <select>
+        <option value = "option-1">Option 1</option>  
+        <option value = "option-2">Option 2</option>  
+        <option value = "option-3">Option 3</option>  
+      </select>
+    );
+}
+
+export default SelectOptions;
