+++ assignment-3/src/utilities/utility.js
@@ -1,0 +1,7 @@
+const newDate = new Date();
+
+const date = newDate.getDate();
+const month = newDate.getMonth() + 1;
+const year = newDate.getFullYear();
+
+export { date, month, year };
